Make Honeypot triggers remove messages from voice channel chats

+++ src/main/java/net/amar/oreojava/events/Honeypot.java
@@ -8,7 +8,9 @@ import net.amar.oreojava.handlers.Verdict;
 import net.dv8tion.jda.api.entities.Guild;
 import net.dv8tion.jda.api.entities.Message;
 import net.dv8tion.jda.api.entities.User;
+
 import net.dv8tion.jda.api.entities.channel.concrete.TextChannel;
+import net.dv8tion.jda.api.entities.channel.concrete.VoiceChannel;
 import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
 import net.dv8tion.jda.api.hooks.ListenerAdapter;
 import org.jetbrains.annotations.NotNull;
@@ -50,11 +52,16 @@ public class Honeypot extends ListenerAdapter {
     }
 
     private void deleteRecentMessagesGuildWide(Guild guild, User user) {
+        deleteVoiceChannelMsgs(guild, user);
+        deleteTextChannelMsgs(guild, user);
+    }
+
+    private void deleteTextChannelMsgs(Guild guild, User user) {
+        List<Message> toDelete = new ArrayList<>();
         OffsetDateTime cutoff = OffsetDateTime.now().minusMinutes(15);
 
         for (TextChannel channel : guild.getTextChannels()) {
             channel.getHistory().retrievePast(100).queue(messages -> {
-                List<Message> toDelete = new ArrayList<>();
 
                 for (Message msg : messages) {
                     if (msg.getTimeCreated().isBefore(cutoff))
@@ -64,8 +71,27 @@ public class Honeypot extends ListenerAdapter {
                         toDelete.add(msg);
                 }
 
-                if (!toDelete.isEmpty())
-                    channel.purgeMessages(toDelete);
+                if (!toDelete.isEmpty()) channel.purgeMessages(toDelete);
+            });
+        }
+    }
+
+    private void deleteVoiceChannelMsgs(Guild guild, User user) {
+        List<Message> toDelete = new ArrayList<>();
+        OffsetDateTime cutoff = OffsetDateTime.now().minusMinutes(15);
+
+        for (VoiceChannel channel : guild.getVoiceChannels()) {
+            channel.getHistory().retrievePast(100).queue(messages -> {
+
+                for (Message msg : messages) {
+                    if (msg.getTimeCreated().isBefore(cutoff))
+                        break;
+
+                    if (msg.getAuthor().getIdLong() == user.getIdLong())
+                        toDelete.add(msg);
+                }
+
+                if (!toDelete.isEmpty()) channel.purgeMessages(toDelete);
             });
         }
     }
